Refactor: Separate style files
